requirements.txt
# Air Quality Agent - Multi-Agent Q&A System

## Dependencies

# Core Framework
langgraph>=0.1.0
langchain>=0.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0

# Database
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# InstructLab
instructlab>=0.5.0
torch>=2.0.0
transformers>=4.30.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# HTTP Clients
aiohttp>=3.8.0
requests>=2.31.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Caching
redis>=4.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring & Logging
structlog>=23.1.0
prometheus-client>=0.17.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3